requirements.txt
python-dotenv
requests
validators
plotly
altair 
pandas 
streamlit